First Commit. Models built and tested.
Models for the database have been created. Two models have been created
apart from the one already existing. Blog model represents the blog which
is owned by some user. Post is the model which represents the content which
is posted by a user and published on a blog. Both blog and post can have
multiple contributors. If a contributor to a blog is set, it automatically
means that it can contribute to all posts inside a blog.

The relationships between the elements are:
1. OneToMany between blog and post
2. OneToMany between user and blog as owner
3. ManyToMany between blog and user as contributors
4. OneToMany between user and post as owner
5. ManyToMany between user and post as contributors