Code Cleanup.
Cleaned up the code so that it is more consistent in conventions, more structured
and more readable. Every view now passes a variable named user to the template
which is used for many purposes. One of the purpose is to provide login/logout
functionality on every page of the blog.